scripts/stress_test/pyproject.toml
[project]
name = "cameraftp-stress-test"
version = "1.0.0"
description = "CameraFTP stress testing tool"
requires-python = ">=3.8"
dependencies = [
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pillow-heif>=0.15.0",
    "piexif>=1.1.0",
]

[tool.uv]
dev-dependencies = []
